backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-allauth
dj-rest-auth
stripe
django-jazzmin
